# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.AssignPermissionsRequest:
    properties:
      permission_ids:
        example:
        - 018f7606-a3f7-7c40-8e4b-2d47c6e04c8d
        items:
          type: string
        minItems: 1
        type: array
    required:
    - permission_ids
    type: object
  github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.CreateRoleRequest:
    properties:
      description:
        example: Can manage operational resources
        maxLength: 255
        type: string
      guard_name:
        example: api
        maxLength: 50
        type: string
      name:
        example: manager
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    type: object
  github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.PermissionResponse:
    properties:
      id:
        example: 018f7606-a3f7-7c40-8e4b-2d47c6e04c8d
        type: string
      name:
        example: users.view
        type: string
    type: object
  github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.RoleResponse:
    properties:
      created_at:
        example: "2026-05-13T10:00:00+07:00"
        type: string
      description:
        example: Can manage operational resources
        type: string
      guard_name:
        example: api
        type: string
      id:
        example: 018f7606-a3f7-7c40-8e4b-2d47c6e04c8d
        type: string
      name:
        example: manager
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.PermissionResponse'
        type: array
      updated_at:
        example: "2026-05-13T10:00:00+07:00"
        type: string
    type: object
  github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.UpdateRoleRequest:
    properties:
      description:
        example: Can manage operational resources
        maxLength: 255
        type: string
      guard_name:
        example: api
        maxLength: 50
        type: string
      name:
        example: manager
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  github_com_ramdhanrizkij_bytecode-api_internal_shared_response.Meta:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: OK
        type: string
    type: object
  github_com_ramdhanrizkij_bytecode-api_internal_shared_response.PaginationMeta:
    properties:
      current_page:
        example: 1
        type: integer
      per_page:
        example: 10
        type: integer
      total_items:
        example: 42
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
  github_com_ramdhanrizkij_bytecode-api_internal_shared_validator.ValidationErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
      tag:
        type: string
    type: object
...
  swaggerdocs.ErrorResponse:
    properties:
      meta:
        $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_shared_response.Meta'
    type: object
...
  swaggerdocs.RoleListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.RoleResponse'
        type: array
      meta:
        $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_shared_response.Meta'
      pagination:
        $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_shared_response.PaginationMeta'
    type: object
  swaggerdocs.RoleResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.RoleResponse'
      meta:
        $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_shared_response.Meta'
    type: object
...
  swaggerdocs.ValidationErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_shared_validator.ValidationErrorDetail'
        type: array
      meta:
        $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_shared_response.Meta'
    type: object
info:
  contact:
    email: support@example.com
    name: Bytecode API Support
  description: REST API for authentication, users, roles, permissions, RBAC, storage-backed
    profiles, and worker-enabled background tasks.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Bytecode API
  version: "1.0"
paths:
...
  /roles/:
    get:
      description: Returns paginated roles with their permissions.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      - default: created_at
        description: Sort field
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: Search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.RoleListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Creates a new role.
      parameters:
      - description: Role payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swaggerdocs.RoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swaggerdocs.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - Roles
  /roles/{id}:
    delete:
      description: Deletes a role by UUID.
      parameters:
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - Roles
    get:
      description: Returns one role by UUID.
      parameters:
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.RoleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Updates an existing role by UUID.
      parameters:
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      - description: Role payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.RoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swaggerdocs.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Roles
  /roles/{id}/permissions:
    delete:
      consumes:
      - application/json
      description: Removes permissions from a role by UUID.
      parameters:
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      - description: Permission IDs
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.AssignPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swaggerdocs.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove permissions from role
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Adds permissions to a role by UUID.
      parameters:
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      - description: Permission IDs
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_ramdhanrizkij_bytecode-api_internal_features_role_domain.AssignPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swaggerdocs.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign permissions to role
      tags:
      - Roles
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"